dev-requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile dev-requirements.in
#
asgiref==3.8.1
    # via
    #   django
    #   django-allauth
    #   django-stubs
black==25.1.0
    # via -r dev-requirements.in
cfgv==3.4.0
    # via pre-commit
click==8.1.8
    # via black
colorama==0.4.6
    # via
    #   click
    #   pytest
coverage[toml]==7.8.0
    # via pytest-cov
distlib==0.3.9
    # via virtualenv
django==4.2.20
    # via
    #   -r F:\social_blog\requirements.in
    #   django-allauth
    #   django-components
    #   django-debug-toolbar
    #   django-stubs
    #   django-stubs-ext
    #   django-tailwind
django-allauth==65.7.0
    # via -r F:\social_blog\requirements.in
django-components==0.138
    # via -r F:\social_blog\requirements.in
django-debug-toolbar==5.1.0
    # via -r dev-requirements.in
django-stubs==5.1.3
    # via -r dev-requirements.in
django-stubs-ext==5.1.3
    # via django-stubs
django-tailwind==4.0.1
    # via -r F:\social_blog\requirements.in
djc-core-html-parser==1.0.2
    # via django-components
filelock==3.18.0
    # via virtualenv
flake8==7.2.0
    # via -r dev-requirements.in
identify==2.6.9
    # via pre-commit
iniconfig==2.1.0
    # via pytest
isort==6.0.1
    # via -r dev-requirements.in
mccabe==0.7.0
    # via flake8
mypy==1.15.0
    # via -r dev-requirements.in
mypy-extensions==1.0.0
    # via
    #   black
    #   mypy
nodeenv==1.9.1
    # via pre-commit
packaging==24.2
    # via
    #   black
    #   pytest
pathspec==0.12.1
    # via black
pillow==11.1.0
    # via -r F:\social_blog\requirements.in
platformdirs==4.3.7
    # via
    #   black
    #   virtualenv
pluggy==1.5.0
    # via pytest
pre-commit==4.2.0
    # via -r dev-requirements.in
psycopg2-binary==2.9.10
    # via -r F:\social_blog\requirements.in
pycodestyle==2.13.0
    # via flake8
pyflakes==3.3.2
    # via flake8
pytest==8.3.5
    # via
    #   -r dev-requirements.in
    #   pytest-cov
    #   pytest-django
pytest-cov==6.1.1
    # via -r dev-requirements.in
pytest-django==4.11.1
    # via -r dev-requirements.in
python-dotenv==1.1.0
    # via -r F:\social_blog\requirements.in
pyyaml==6.0.2
    # via pre-commit
sqlparse==0.5.3
    # via
    #   django
    #   django-debug-toolbar
types-pyyaml==6.0.12.20250402
    # via django-stubs
typing-extensions==4.13.2
    # via
    #   django-stubs
    #   django-stubs-ext
    #   mypy
tzdata==2025.2
    # via django
virtualenv==20.30.0
    # via pre-commit
